second

+++ models.py
@@ -49,11 +49,26 @@ async def get_async_session(
     return async_session_maker
 
 
-async def get_async_session():
+async def get_async_session(
+    drop: bool = False, create: bool = False
+):
+
     async with engine.begin() as conn:
-        await conn.run_sync(Base.metadata.drop_all)
-        await conn.run_sync(Base.metadata.create_all)
+        if drop:
+            await conn.run_sync(Base.metadata.drop_all)
+        if create:
+            print(1)
+            await conn.run_sync(Base.metadata.create_all)
+    async_session_maker = sessionmaker(
+        engine, expire_on_commit=False, class_=AsyncSession
+    )
+
+    return async_session_maker
+
+
+async def models_main():
+    await get_async_session(True, True)
 
 
 if __name__ == '__main__':
-    asyncio.run(get_async_session())
+    asyncio.run(models_main())
